Move Editor class to new Back_Compat directory
