+++ src/Events/Classy/Controller.php
@@ -11,6 +11,7 @@ namespace TEC\Events\Classy;
 
 use TEC\Common\Contracts\Provider\Controller as ControllerContract;
 use TEC\Common\StellarWP\Assets\Asset;
+use TEC\Events\Classy\Back_Compat\Editor;
 use TEC\Events\Custom_Tables\V1\Models\Event;
 use TEC\Events\Custom_Tables\V1\Models\Occurrence;
 use Tribe__Events__Main as TEC;
@@ -145,7 +146,7 @@ class Controller extends ControllerContract {
 	 */
 	protected function do_register(): void {
 		// Register the `editor` binding replacement for back-compatibility purposes.
-		$back_compatible_editor = new Back_Compatible_Editor();
+		$back_compatible_editor = new Editor();
 		$this->container->singleton( 'editor', $back_compatible_editor );
 		$this->container->singleton( 'events.editor', $back_compatible_editor );
 		$this->container->singleton( 'events.editor.compatibility', $back_compatible_editor );
